Use integer ids instead of doubles

+++ LuxMob/ScriptMessageHandler.swift
@@ -13,7 +13,7 @@ typealias Action = [String: Any]
 
 
 struct Response: Codable {
-    var id: Double
+    var id: Int
     var command: String
     var args: [String?]
     var msg: String?
@@ -126,7 +126,7 @@ class ScriptMessageHandler: NSObject, WKScriptMessageHandler {
 }
 
     private func postResponseToWebview(args: [String?], _ action: Action) {
-        let id = action["id"] as! Double
+        let id = action["id"] as! Int
         let command = "response"
         let response: Response = Response(id: id, command: command, args: args, msg: "")
         let data = try! encoder.encode(response)
@@ -134,7 +134,7 @@ class ScriptMessageHandler: NSObject, WKScriptMessageHandler {
     }
     
     private func postErrorToWebview(msg: String, _ action: Action) {
-        let id = action["id"] as! Double
+        let id = action["id"] as! Int
         let response: Response = Response(id: id, command: "error", args: [], msg: msg)
         let data = try! encoder.encode(response)
         postDataToWebview(data)
